+++ python/latexminted/cmdline.py
@@ -19,10 +19,11 @@ from latex2pydata import __version__ as latex2pydata_version
 from pygments import __version__ as pygments_version
 from .version import __version__
 from .command_batch import batch
-from .command_clean import clean, clean_file, clean_messages
+from .command_clean import clean, clean_file, clean_messages, paths_skipped_in_initial_temp_cleaning
 from .command_config import config
 from .command_highlight import highlight
 from .command_styledef import styledef
+from .debug import debug_mv_data
 from .load_data import load_data
 from .messages import Messages
 
@@ -43,35 +44,41 @@ def main():
     parser_batch = subparsers.add_parser('batch', help='Batch process highlight, styledef, and clean')
     parser_batch.set_defaults(func=batch)
     parser_batch.add_argument('--timestamp', help='LaTeX compile timestamp', required=True)
+    parser_batch.add_argument('--debug', help='Keep temp files for debugging', action='store_true')
     parser_batch.add_argument('md5', help=r'MD5 hash based \jobname')
 
     parser_clean = subparsers.add_parser('clean', help='Clean up temp files and unused cache files')
     parser_clean.set_defaults(func=clean)
     parser_clean.add_argument('--timestamp', help='LaTeX compile timestamp', required=True)
+    parser_clean.add_argument('--debug', help='Keep temp files for debugging', action='store_true')
     parser_clean.add_argument('md5', help=r'MD5 hash based \jobname')
 
     parser_cleanfile = subparsers.add_parser('cleanfile', help='Clean up a particular temp file')
     parser_cleanfile.set_defaults(func=clean_file)
+    parser_cleanfile.add_argument('--debug', help='Keep temp files for debugging', action='store_true')
     parser_cleanfile.add_argument('file', help='File to clean')
 
     parser_config = subparsers.add_parser('config', help='Detect configuration and save it to file for \\input')
     parser_config.set_defaults(func=config)
     parser_config.add_argument('--timestamp', help='LaTeX compile timestamp', required=True)
+    parser_config.add_argument('--debug', help='Keep temp files for debugging', action='store_true')
     parser_config.add_argument('md5', help=r'MD5 hash based \jobname')
 
     parser_highlight = subparsers.add_parser('highlight', help='Highlight code and save it to file for \\input')
     parser_highlight.set_defaults(func=highlight)
     parser_highlight.add_argument('--timestamp', help='LaTeX compile timestamp', required=True)
+    parser_highlight.add_argument('--debug', help='Keep temp files for debugging', action='store_true')
     parser_highlight.add_argument('md5', help=r'MD5 hash based \jobname')
 
     parser_styledef = subparsers.add_parser('styledef', help='Generate highlighting style definition and save it to file for \\input')
     parser_styledef.set_defaults(func=styledef)
     parser_styledef.add_argument('--timestamp', help='LaTeX compile timestamp', required=True)
+    parser_styledef.add_argument('--debug', help='Keep temp files for debugging', action='store_true')
     parser_styledef.add_argument('md5', help=r'MD5 hash based \jobname')
 
     cmdline_args = parser.parse_args()
 
-    func_keys = set(['md5', 'file', 'timestamp'])
+    func_keys = set(['md5', 'file', 'timestamp', 'debug'])
     func_args = {k: v for k, v in vars(cmdline_args).items() if k in func_keys}
     md5: str | None = func_args.get('md5')
     timestamp: str | None = func_args.get('timestamp')
@@ -83,7 +90,7 @@ def main():
     messages = Messages(md5=md5)
     func_args['messages'] = messages
     try:
-        data = load_data(md5=md5, messages=messages, timestamp=timestamp, command=cmdline_args.subparser_name)
+        maybe_data = load_data(md5=md5, messages=messages, timestamp=timestamp, command=cmdline_args.subparser_name)
     except Exception as e:
         messages.append_error(
             rf'Failed due to unexpected error (see \detokenize{{"{messages.errlog_file_name}"}} if it exists)'
@@ -98,13 +105,19 @@ def main():
     if messages.has_errors():
         messages.communicate()
         sys.exit(1)
-    if data is None:
+    if maybe_data is None:
         messages.append_error('Unexpectedly received no data without any error messages')
         messages.communicate()
         sys.exit(1)
 
+    data, data_path = maybe_data
     func_args['data'] = data
+    debug: bool = func_args.get('debug', False)
+    if debug:
+        paths_skipped_in_initial_temp_cleaning.add(data_path)
     cmdline_args.func(**func_args)
+    if debug:
+        debug_mv_data(md5=md5, data_path=data_path)
     messages.communicate()
     if messages.has_errors():
         sys.exit(1)