+++ python/latexminted/debug.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2024, Geoffrey M. Poore
+# All rights reserved.
+#
+# Licensed under the LaTeX Project Public License version 1.3c:
+# https://www.latex-project.org/lppl.txt
+#
+
+
+from __future__ import annotations
+
+from .err import PathSecurityError
+from .restricted import RestrictedPath
+
+
+
+
+def debug_mv_data(*, md5, data_path: RestrictedPath) -> RestrictedPath | None:
+    for n in range(1, 101):
+        replacement_path = data_path.parent / f'_{md5}_{n}.data.minted'
+        if not replacement_path.is_file():
+            break
+    else:
+        return None
+    try:
+        data_path.replace(replacement_path)
+    except (FileNotFoundError, PermissionError, PathSecurityError):
+        return None
+    return replacement_path